01_calendar_reminder/backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
apscheduler
pytz
requests 